+++ start/src/global.d.ts
@@ -84,6 +84,7 @@ declare module 'vue' {
     FInfiniteScrolling: typeof components.FInfiniteScrolling
     FNumberAnimate: typeof components.FNumberAnimate
     FCountDown: typeof components.FCountDown
+    FAvatarGroup: typeof components.FAvatarGroup
   }
 
   interface ComponentCustomProperties {